requirements.txt
kafka-python
python-dotenv
boto3
pyspark==3.5.0
